+++ gen/templates/feature_templates/data_source.go
@@ -32,6 +32,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/go-sdwan"
 	"github.com/CiscoDevNet/terraform-provider-sdwan/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
 )
 
 // Ensure the implementation satisfies the expected interfaces.
@@ -60,7 +61,8 @@ func (d *{{camelCase .Name}}FeatureTemplateDataSource) Schema(ctx context.Contex
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				MarkdownDescription: "The id of the feature template",
-				Required:            true,
+				Optional:            true,
+				Computed:            true,
 			},
 			"version": schema.Int64Attribute{
 				MarkdownDescription: "The version of the feature template",
@@ -72,6 +74,7 @@ func (d *{{camelCase .Name}}FeatureTemplateDataSource) Schema(ctx context.Contex
 			},
 			"name": schema.StringAttribute{
 				MarkdownDescription: "The name of the feature template",
+				Optional:            true,
 				Computed:            true,
 			},
 			"description": schema.StringAttribute{
@@ -177,6 +180,16 @@ func (d *{{camelCase .Name}}FeatureTemplateDataSource) Schema(ctx context.Contex
 	}
 }
 
+func (d *{{camelCase .Name}}FeatureTemplateDataSource) ConfigValidators(ctx context.Context) []datasource.ConfigValidator {
+    return []datasource.ConfigValidator{
+        datasourcevalidator.ExactlyOneOf(
+            path.MatchRoot("id"),
+            path.MatchRoot("name"),
+        ),
+    }
+}
+
+
 func (d *{{camelCase .Name}}FeatureTemplateDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
@@ -196,6 +209,28 @@ func (d *{{camelCase .Name}}FeatureTemplateDataSource) Read(ctx context.Context,
 	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", config.Id.String()))
+
+	if config.Id.IsNull() && !config.Name.IsNull() {
+		res, err := d.client.Get("/template/feature")
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve existing templates, got error: %s", err))
+			return
+		}
+		if value := res.Get("data"); len(value.Array()) > 0 {
+			value.ForEach(func(k, v gjson.Result) bool {
+				if config.Name.ValueString() == v.Get("templateName").String() {
+					config.Id = types.StringValue(v.Get("templateId").String())
+					tflog.Debug(ctx, fmt.Sprintf("%s: Found feature template with name '%v', id: %v", config.Id.String(), config.Name.ValueString(), config.Id.String()))
+					return false
+				}
+				return true
+			})
+		}
+		if config.Id.IsNull() {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to find feature template with name: %s", config.Name.ValueString()))
+			return
+		}
+	}
 
 	res, err := d.client.Get("/template/feature/object/" + config.Id.ValueString())
 	if err != nil {