+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.2.7 (unreleased)
+
+- Add option to feature template data sources to identify templates by its name
+
 ## 0.2.6
 
 - Fix deployment of attached device templates without feature template changes